refactor: move server dispatch into dedicated module
